requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==3.2.7
django-bootstrap4==3.0.1
django-chartjs==2.2.1
pytz==2021.1
soupsieve==2.2.1
sqlparse==0.4.2
